src/main/java/com/polixis/batchprocessing/PersonItemProcessor.java
package com.polixis.batchprocessing;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.batch.item.ItemProcessor;

public class PersonItemProcessor implements ItemProcessor<PersonDataSet, PersonDataSet> {

	private static final Logger log = LoggerFactory.getLogger(PersonItemProcessor.class);

	@Override
	public PersonDataSet process(final PersonDataSet person) {
		final String firstName = person.getFirstName().toUpperCase();
		final String lastName = person.getLastName().toUpperCase();
		final String date = person.getDate().toUpperCase();

		final PersonDataSet transformedPerson = new PersonDataSet(firstName, lastName, date);

		log.info("Converting (" + person + ") into (" + transformedPerson + ")");

		return transformedPerson;
	}

}
